requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-Mail==0.9.1
Flask-JWT-Extended==4.3.1
Werkzeug==2.0.1
python-dotenv==0.19.0
PyMySQL==1.0.2
cryptography==3.4.7
mysqlclient==2.2.0
Pillow==10.0.0
gunicorn==21.2.0 